webapp/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0
